text size math

+++ packages/extension-text/src/components/TextDragStart.ts
@@ -2,10 +2,10 @@ import { component, field } from '@lastolivegames/becsy'
 
 @component
 export class TextDragStart {
-  @field.float32 public declare startHeight: number
-  @field.float32 public declare startWidth: number
-  @field.float32 public declare startFontSize: number
+  @field.float64 public declare startHeight: number
+  @field.float64 public declare startWidth: number
+  @field.float64 public declare startFontSize: number
 
-  @field.float32 public declare lastHeight: number
-  @field.float32 public declare lastWidth: number
+  @field.float64 public declare lastHeight: number
+  @field.float64 public declare lastWidth: number
 }
